Implemented - Options Menu + Player/Enemy Starts
This commit implements some simple functionality for the Options Menu and enables Enemy & Player positions to be enabled on NewGame().

References to the Player & Enemy are also stored in StaticManagment.

Next, implement the basic options the player will have for the game. Decide if it will be idle, auto, or actual in-depth combat.