Implemented - UI Changes/Script Updates + Enemy Spawning
This commit adds a bunch of changes to the UI including buttons for the player to use Spells & Abilities. It also adds a couple prefab variants of some enemies that have been used to test the new spawning mechanic.

Next, implement ways for the player and enemy to interact with each other.