Implemented - Basic Magic System + FireSpark Spell
This commit implements the first parts of a basic magic system that let's the player spend mana from their mana pool on spells that apply either damage or later on a range of extra effects.

Next, continue work on this magic system and the overall design of the game's systems.