Implemented - Lightning & Water Magic Scripts
This commit adds two new magic types, lightning and water. Each one can be cast by the player the same way as fire magic, with a mana cost.

Next, play some of that idle phone game to get some inspiration as to how to make the use of magic more fluid/fun.