Fixed - PetInfoBox Text Scalling + Spellcool Down Menu Issue
This commit fixes a couple different issues. The first being the text scalling for the PetInfoBox Name field, before it was made way too small. The other issue was that if the player used a spell and closed the spell menu it would break the cooldown, preventing the player from using the spell again.

Next, start work on allowing the player to send out and recall their pets to specific positions.