Improved - Pet Info Box UI Layout & Sprite Naming
This commit improves the UI of the PetInfoBox section, the Pet name and DPS info is better layed out and never displays the (Clone) part of the name anymore.

The issue with the Recall/Deploy that I suggested be fixed next was actually a non-issue and a miscommunication on my part.

Next, allow the deployed pets to deal idle damage based on their DPS stat.