Implemented - DamageNumbers Fade Out
This commit implements functionality for the DamageNumbers to fade out over time, this makes them look cleaner and more fitting with the rest of the game.

Next, implement a way to keep track of how many 'Stages' the player has completed. This will eventually be used to change the type of enemy that is spawned. Similar to Tap Titans 1/2.